feat: period create method

+++ src/main/kotlin/kr/picktime/timetable/period/domain/entity/PeriodEntity.kt
@@ -10,7 +10,7 @@ data class PeriodEntity(
     @Id
     val id: Long? = null,
     val schoolId: Long,
-    val period: Int,
+    val period: Long,
     val startTime: LocalTime,
     val endTime: LocalTime,
     val dayOfWeek: DayOfWeek,
